WW-4954 Moves XWorkList out of util package
